Fix SA1402 violation in PPMouseLegacy.cs
