+++ PowerPointLabs/PowerPointLabs/PowerPointLabs.csproj
@@ -626,7 +626,10 @@
     <Compile Include="PositionsLab\ReorderSettingsDialog.xaml.cs">
       <DependentUpon>ReorderSettingsDialog.xaml</DependentUpon>
     </Compile>
-    <Compile Include="PPMouseLegacy.cs" />
+    <Compile Include="PPExtraEventHelper\LMouseDownListener.cs" />
+    <Compile Include="PPExtraEventHelper\LMouseUpListener.cs" />
+    <Compile Include="PPExtraEventHelper\SysMouseEventInfo.cs" />
+    <Compile Include="PPExtraEventHelper\PPMouseLegacy.cs" />
     <Compile Include="ResizeLab\AdjustProportionallySettingsDialog.xaml.cs">
       <DependentUpon>AdjustProportionallySettingsDialog.xaml</DependentUpon>
     </Compile>
